+++ aql/src/main/java/org/ehrbase/openehr/sdk/aql/dto/operand/TemporalPrimitive.java
@@ -19,20 +19,39 @@ package org.ehrbase.openehr.sdk.aql.dto.operand;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.time.temporal.ChronoField;
 import java.time.temporal.TemporalAccessor;
-import org.apache.commons.lang3.StringUtils;
+import java.util.regex.Pattern;
 import org.ehrbase.openehr.sdk.util.OpenEHRDateTimeParseUtils;
-import org.ehrbase.openehr.sdk.util.OpenEHRDateTimeSerializationUtils;
+import org.ehrbase.openehr.sdk.util.OpenEhrTemporal;
+import org.ehrbase.openehr.sdk.util.RegexUtils;
 
 /**
  * @author Stefan Spiska
  */
 public final class TemporalPrimitive extends StringPrimitive {
+    private static final Pattern TIME_REGEX;
+
+    static {
+        String HOUR = RegexUtils.nonCapturing(RegexUtils.or("[01][0-9]", "[2][0-3]"));
+        String MINUTE = "[0-5][0-9]";
+        String SECOND = MINUTE;
+        String TIME_SHORT = HOUR + MINUTE + SECOND;
+        String TIME_LONG = HOUR + ':' + MINUTE + ':' + SECOND;
+        String TIMEZONE =
+                RegexUtils.or("Z", RegexUtils.nonCapturing("[-+]", HOUR, RegexUtils.optional("[:]?" + MINUTE)));
+        String MILLISECOND = ".[0-9][0-9][0-9]";
+        TIME_REGEX = Pattern.compile("^" + RegexUtils.or(TIME_SHORT, TIME_LONG) + RegexUtils.optional(MILLISECOND)
+                + RegexUtils.optional(TIMEZONE) + "$");
+    }
 
     private TemporalAccessor temporal;
 
     public TemporalPrimitive() {}
+
+    private TemporalPrimitive(String value, TemporalAccessor temporal) {
+        super(value);
+        this.temporal = temporal;
+    }
 
     public static StringPrimitive fromString(String value) {
         TemporalAccessor temporal;
@@ -42,10 +61,7 @@ public final class TemporalPrimitive extends StringPrimitive {
             return new StringPrimitive(value);
         }
 
-        TemporalPrimitive temporalPrimitive = new TemporalPrimitive();
-        temporalPrimitive.setTemporal(temporal);
-
-        return temporalPrimitive;
+        return new TemporalPrimitive(value, temporal);
     }
 
     @JsonProperty
@@ -59,13 +75,14 @@ public final class TemporalPrimitive extends StringPrimitive {
         if (value == null) {
             return null;
         }
-        if (StringUtils.containsAny(value, 'T', '-')) {
-            return OpenEHRDateTimeParseUtils.parseDateTime(value);
-        }
-        if (StringUtils.containsAny(value, ':', '.', '+', '-', 'Z') || value.length() == 6) {
+        if (TIME_REGEX.matcher(value).matches()) {
             return OpenEHRDateTimeParseUtils.parseTime(value);
         }
-        return OpenEHRDateTimeParseUtils.parseDateTime(value);
+        if (value.indexOf('T') >= 0) {
+            return OpenEHRDateTimeParseUtils.parseDateTime(value);
+        } else {
+            return OpenEHRDateTimeParseUtils.parseDate(value);
+        }
     }
 
     @JsonIgnore
@@ -74,14 +91,12 @@ public final class TemporalPrimitive extends StringPrimitive {
     }
 
     public void setTemporal(TemporalAccessor temporal) {
-        this.temporal = temporal;
         if (temporal == null) {
-            setValue(null);
+            this.temporal = null;
+            super.setValue(null);
         } else {
-            setValue(
-                    temporal.isSupported(ChronoField.YEAR)
-                            ? OpenEHRDateTimeSerializationUtils.formatDateTime(temporal)
-                            : OpenEHRDateTimeSerializationUtils.formatTime(temporal));
+            this.temporal = new OpenEhrTemporal(temporal);
+            super.setValue(this.temporal.toString());
         }
     }