+++ test-data/src/main/java/org/ehrbase/openehr/sdk/test_data/operationaltemplate/OperationalTemplateTestData.java
@@ -113,7 +113,11 @@ public enum OperationalTemplateTestData {
             "EHRN-ABDM-OPConsultRecord.v2.0"),
     //  TEST_ISM("ISM transition test", "test-ism.vitagroup.de.v1.opt", "test-ism.vitagroup.de.v1");
     INFORME_AMB_1_ARQUETIP_OBS("episodic example", "informe_amb_1_arquetip_OBS.opt", "informe_amb_1_arquetip_OBS"),
-    SPANISH_EXAMPLE("Spanish example", "my_spanish_template_v0.opt", "my_spanish_template_v0");
+    SPANISH_EXAMPLE("Spanish example", "my_spanish_template_v0.opt", "my_spanish_template_v0"),
+    INTERVAL_PARTIAL_DATE(
+            "Interval with partial date time constraint",
+            "Test_quantity_dv_interval_dv_date_time_open_constraint.v0.opt",
+            "Test_quantity_dv_interval_dv_date_time_open_constraint.v0");
 
     public static final String OPERATIONALTEMPLATE_PATH_SEGMENT = "operationaltemplate";
     private final String filename;