+++ CHANGELOG.md
@@ -3,12 +3,33 @@
 Note: version releases in the 0.x.y range may introduce breaking changes.
 
 ## [unreleased]
+ ### Added 
+ ### Fixed 
+
+## [2.33.1]
+ ### Added 
+ ### Fixed 
+ - Retain original string when parsing temporals in AQL [#752](https://github.com/ehrbase/openEHR_SDK/pull/752)
+
+## [2.33.0]
+ ### Added 
+ ### Fixed 
+ - Validation of DV_CODED_TEXT.value for template-defined terminologies [#750](https://github.com/ehrbase/openEHR_SDK/pull/750)
+
+## [2.32.0]
  ### Added
+- Support for partial time [#747](https://github.dev/ehrbase/openEHR_SDK/pull/747)
  ### Changed
  ### Fixed
-- Fixed terminology validation of DV_ORDINAL.value and DV_CODED_TEXT.value [#739](https://github.com/ehrbase/openEHR_SDK/pull/739)
-- Fixed terminology validation for e.g. Spanish templates [#735](https://github.com/ehrbase/openEHR_SDK/pull/735)
+- Terminology validation of DV_ORDINAL.value and DV_CODED_TEXT.value [#739](https://github.com/ehrbase/openEHR_SDK/pull/739)
+- Terminology validation for e.g. Spanish templates [#735](https://github.com/ehrbase/openEHR_SDK/pull/735)
 - Regardless of the composition language, the English terms of openehr terminologies are now always accepted [#735](https://github.com/ehrbase/openEHR_SDK/pull/735)
+- Validation of limits of intervals with partial dates [#747](https://github.dev/ehrbase/openEHR_SDK/pull/747)
+- Prevent mixing of long and short time format [#747](https://github.dev/ehrbase/openEHR_SDK/pull/747)
+
+## [2.31.1]
+ ### Fixed
+- Retain original string when parsing temporals in AQL (backport) [#752](https://github.com/ehrbase/openEHR_SDK/pull/752)
 
 ## [2.31.0]
  ### Added 
@@ -542,4 +563,7 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 [2.29.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.28.0...v2.29.0
 [2.30.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.29.0...v2.30.0
 [2.31.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.30.0...v2.31.0
-[unreleased]: https://github.com/ehrbase/openEHR_SDK/compare/v2.31.0...HEAD
+[2.32.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.31.0...v2.32.0
+[2.33.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.32.0...v2.33.0
+[2.33.1]: https://github.com/ehrbase/openEHR_SDK/compare/v2.33.0...v2.33.1
+[unreleased]: https://github.com/ehrbase/openEHR_SDK/compare/v2.33.1...HEAD